test(Sheet): consolidate unit test suites
