--- src/components/Sheet/__tests__/Sheet.swipe.test.tsx
@@ -1,147 +1,0 @@
-import * as React from 'react';
-
-import {act, fireEvent, render, screen} from '../../../../test-utils/utils';
-import {Sheet} from '../Sheet';
-import {SHEET_TRANSITION_DURATION_MS, SheetQa} from '../constants';
-
-const HIDE_THRESHOLD = 50;
-const SHEET_HEIGHT = 300;
-const TOUCH_START_POINT = 100;
-
-describe('Sheet swipe area', () => {
-    let getBoundingClientRectSpy: jest.SpyInstance;
-    beforeEach(() => {
-        jest.useFakeTimers();
-        getBoundingClientRectSpy = jest
-            .spyOn(HTMLElement.prototype, 'getBoundingClientRect')
-            .mockReturnValue({height: SHEET_HEIGHT, width: 0, top: 0, left: 0} as DOMRect);
-    });
-    afterEach(() => {
-        getBoundingClientRectSpy.mockRestore();
-        jest.clearAllTimers();
-        jest.useRealTimers();
-    });
-
-    function swipeOnArea(area: Element, {from, to}: {from: number; to: number}) {
-        fireEvent.touchStart(area, {touches: [{clientX: 0, clientY: from}]});
-        fireEvent.touchMove(area, {touches: [{clientX: 0, clientY: to}]});
-        fireEvent.touchEnd(area, {touches: [{clientX: 0, clientY: to}]});
-    }
-
-    test('closes the sheet only when the swipe-area movement exceeds the threshold', () => {
-        const onClose = jest.fn();
-        const onOpenChange = jest.fn();
-
-        function AcceptingSheet() {
-            const [visible, setVisible] = React.useState(true);
-
-            return (
-                <Sheet
-                    visible={visible}
-                    onClose={onClose}
-                    onOpenChange={(open, event, reason) => {
-                        onOpenChange(open, event, reason);
-                        setVisible(open);
-                    }}
-                >
-                    Content
-                </Sheet>
-            );
-        }
-
-        render(<AcceptingSheet />);
-
-        const swipeArea = screen.getByTestId(SheetQa.SWIPE_AREA);
-        const sheet = screen.getByRole('dialog');
-        const veil = screen.getByTestId(SheetQa.VEIL);
-
-        // A short swipe (below the threshold) returns the sheet to the open position.
-        swipeOnArea(swipeArea, {
-            from: TOUCH_START_POINT,
-            to: TOUCH_START_POINT + (HIDE_THRESHOLD - 20),
-        });
-
-        expect(onClose).not.toHaveBeenCalled();
-        expect(onOpenChange).not.toHaveBeenCalled();
-
-        // show() restores the fully opened state: full veil opacity and shifted-up transform.
-        expect(veil.style.opacity).toBe('1');
-        expect(sheet.style.transform).toBe(`translate3d(0, -${SHEET_HEIGHT}px, 0)`);
-
-        // A swipe above the threshold starts the closing animation.
-        swipeOnArea(swipeArea, {
-            from: TOUCH_START_POINT,
-            to: TOUCH_START_POINT + (HIDE_THRESHOLD + 20),
-        });
-
-        // hide() moves the sheet down and fades the veil out.
-        expect(sheet.style.transform).toBe('translate3d(0, 0, 0)');
-        expect(veil.style.opacity).toBe('0');
-        // onClose fires only after the hiding transition finishes.
-        expect(onClose).not.toHaveBeenCalled();
-        expect(onOpenChange).toHaveBeenCalledWith(false, expect.any(Event), 'swipe');
-        expect(onOpenChange).toHaveBeenCalledTimes(1);
-
-        act(() => {
-            jest.advanceTimersByTime(SHEET_TRANSITION_DURATION_MS);
-        });
-        expect(onClose).toHaveBeenCalledTimes(1);
-    });
-
-    test('keeps the sheet open after a full-height swipe when controlled dismissal is not accepted', () => {
-        const onOpenChange = jest.fn();
-        render(
-            <Sheet visible onOpenChange={onOpenChange}>
-                Content
-            </Sheet>,
-        );
-
-        const swipeArea = screen.getByTestId(SheetQa.SWIPE_AREA);
-        const sheet = screen.getByRole('dialog');
-        const veil = screen.getByTestId(SheetQa.VEIL);
-
-        fireEvent.transitionEnd(veil);
-        swipeOnArea(swipeArea, {
-            from: TOUCH_START_POINT,
-            to: TOUCH_START_POINT + SHEET_HEIGHT,
-        });
-
-        expect(onOpenChange).toHaveBeenCalledWith(false, expect.any(Event), 'swipe');
-        expect(onOpenChange).toHaveBeenCalledTimes(1);
-        expect(veil).toHaveStyle({opacity: '1'});
-        expect(sheet.style.transform).toBe(`translate3d(0, -${SHEET_HEIGHT}px, 0)`);
-        expect(screen.getByRole('dialog')).toBeInTheDocument();
-    });
-
-    test('restores open state after swipe-area touchcancel', () => {
-        const onClose = jest.fn();
-        const onOpenChange = jest.fn();
-
-        render(
-            <Sheet visible onClose={onClose} onOpenChange={onOpenChange}>
-                Content
-            </Sheet>,
-        );
-
-        const swipeArea = screen.getByTestId(SheetQa.SWIPE_AREA);
-        const contentArea = screen.getByTestId(SheetQa.CONTENT_AREA);
-        const sheet = screen.getByRole('dialog');
-        const veil = screen.getByTestId(SheetQa.VEIL);
-
-        fireEvent.touchStart(swipeArea, {touches: [{clientX: 0, clientY: 100}]});
-        fireEvent.touchMove(swipeArea, {touches: [{clientX: 0, clientY: 170}]});
-
-        expect(sheet.style.transform).toBe('translate3d(0, -230px, 0)');
-        expect(contentArea).toHaveClass('g-sheet-content-area_without-scroll');
-
-        fireEvent.touchCancel(swipeArea);
-
-        expect(sheet.style.transform).toBe(`translate3d(0, -${SHEET_HEIGHT}px, 0)`);
-        expect(veil).toHaveStyle({opacity: '1'});
-        expect(sheet).toHaveClass('g-sheet__sheet_with-transition');
-        expect(veil).toHaveClass('g-sheet-veil_with-transition');
-        expect(contentArea).not.toHaveClass('g-sheet-content-area_without-scroll');
-        expect(onOpenChange).not.toHaveBeenCalled();
-        expect(onClose).not.toHaveBeenCalled();
-    });
-});
